requirements.txt
Pillow==10.3.0
streamlit==1.33.0
ultralytics==8.1.43
